fix: Update drag and drop component and behavior
