+++ src/components/home/tasks/Tasks.tsx
@@ -6,56 +6,16 @@ import {
   type Task,
 } from "../../../generated/graphql";
 import useCardStore from "../../../store/useEditManager";
-import { useMemo, useState, useEffect, useRef } from "react";
+import { useMemo, useEffect } from "react";
 import { columnName } from "../../../hooks/columnName";
 import { LoadingComponent } from "../../ui/loading/Loading";
-import {
-  DndContext,
-  KeyboardSensor,
-  PointerSensor,
-  useSensor,
-  useSensors,
-  type DragEndEvent,
-  type DragStartEvent,
-  type DragOverEvent,
-  DragOverlay,
-  rectIntersection,
-} from "@dnd-kit/core";
-import { arrayMove } from "@dnd-kit/sortable";
-import { useMutation } from "@apollo/client";
-import { EDIT_TASK } from "../../../api/graphql/queries.graphql";
+import { DndContext, DragOverlay, rectIntersection } from "@dnd-kit/core";
 import { TaskColumn } from "../taskColumns/TaskColumns";
 import { Cards } from "../cards/Cards";
-import { useCustomToast } from "../../../hooks/UseCustomToast";
+import { useDragAndDrop } from "../../../hooks/useDragAndDrop";
 
 export const Tasks = () => {
-  const { showToast } = useCustomToast();
   const { searchCardElement, filtersElement } = useCardStore();
-  const [activeTask, setActiveTask] = useState<Task | null>(null);
-  const [isDragging, setIsDragging] = useState(false);
-  const [optimisticTasks, setOptimisticTasks] = useState<Task[]>([]);
-  const [pendingUpdates, setPendingUpdates] = useState<Set<string>>(new Set());
-
-  const pendingUpdatesRef = useRef<Set<string>>(new Set());
-  const prevSnapshotRef = useRef<Task[] | null>(null);
-
-  const addPending = (id: string) => {
-    pendingUpdatesRef.current.add(id);
-    setPendingUpdates(new Set(pendingUpdatesRef.current));
-  };
-  const removePending = (id: string) => {
-    pendingUpdatesRef.current.delete(id);
-    setPendingUpdates(new Set(pendingUpdatesRef.current));
-  };
-  const beginUpdateSnapshot = () => {
-    prevSnapshotRef.current = optimisticTasks;
-  };
-  const revertIfNeeded = () => {
-    if (prevSnapshotRef.current) {
-      setOptimisticTasks(prevSnapshotRef.current);
-      prevSnapshotRef.current = null;
-    }
-  };
 
   const queryInput = useMemo(() => {
     const input: FilterTaskInput = {};
@@ -85,8 +45,6 @@ export const Tasks = () => {
     return input;
   }, [searchCardElement, filtersElement]);
 
-  const lastTaskIdRef = useRef<string | null>(null);
-
   const { loading: isloadingCols, error: errorCols } = useGetColumnsQuery();
   const {
     loading: isloadingTasks,
@@ -102,23 +60,42 @@ export const Tasks = () => {
     errorPolicy: "all",
   });
 
+  const tasksToDisplay = useMemo(() => {
+    const serverTasks = tasksData?.tasks || [];
+    return serverTasks as Task[];
+  }, [tasksData?.tasks]);
+
+  const {
+    activeTask,
+    isDragging,
+    optimisticTasks,
+    pendingUpdates,
+    sensors,
+    handleDragStart,
+    handleDragOver,
+    handleDragEnd,
+    setOptimisticTasks,
+  } = useDragAndDrop({
+    tasks: tasksToDisplay,
+    tasksData: tasksData ? { tasks: tasksToDisplay } : undefined,
+  });
+
   useEffect(() => {
     if (tasksData?.tasks) {
       setOptimisticTasks(tasksData.tasks as Task[]);
     }
-  }, [tasksData?.tasks]);
+  }, [tasksData?.tasks, setOptimisticTasks]);
 
-  const tasksToDisplay = useMemo(() => {
+  const finalTasksToDisplay = useMemo(() => {
     if (pendingUpdates.size > 0 || isDragging) {
       return optimisticTasks;
     }
-    const serverTasks = tasksData?.tasks || [];
-    return serverTasks as Task[];
-  }, [pendingUpdates.size, isDragging, optimisticTasks, tasksData?.tasks]);
+    return tasksToDisplay;
+  }, [pendingUpdates.size, isDragging, optimisticTasks, tasksToDisplay]);
 
-  const tasksByStatus = useMemo(
+  const finalTasksByStatus = useMemo(
     () =>
-      tasksToDisplay?.reduce(
+      finalTasksToDisplay?.reduce(
         (acc, task) => {
           acc[task.status] = acc[task.status]
             ? [...acc[task.status], task]
@@ -127,189 +104,12 @@ export const Tasks = () => {
         },
         {} as Record<string, Task[]>,
       ) || {},
-    [tasksToDisplay],
+    [finalTasksToDisplay],
   );
 
   const orderedColumns = Object.keys(columnName).map((key) => ({
     name: key,
   }));
-
-  const [updateTask] = useMutation(EDIT_TASK, {
-    onError: (error) => {
-      console.error("Error updating task:", error);
-      pendingUpdatesRef.current.clear();
-      setPendingUpdates(new Set());
-      revertIfNeeded();
-      setIsDragging(false);
-      showToast("error", "the card won't be updated");
-    },
-    onCompleted: () => {
-      if (lastTaskIdRef.current) {
-        removePending(lastTaskIdRef.current);
-        lastTaskIdRef.current = null;
-      }
-      setIsDragging(false);
-      showToast("success", "the card was updated successfully");
-    },
-  });
-
-  const applyLocalStatusChange = (
-    taskId: string,
-    status: Task["status"],
-  ): void => {
-    setOptimisticTasks((prev) =>
-      prev.map((t) => (t.id === taskId ? { ...t, status } : t)),
-    );
-  };
-
-  function handleDragStart(event: DragStartEvent) {
-    const taskIdRaw = event.active.id as string;
-    const taskId = taskIdRaw.replace("task:", "");
-
-    const currentTask = tasksToDisplay.find((task) => task.id === taskId);
-    setActiveTask(currentTask || null);
-    setIsDragging(true);
-
-    if (optimisticTasks.length !== tasksToDisplay.length) {
-      setOptimisticTasks([...tasksToDisplay]);
-    }
-  }
-
-  function handleDragOver(event: DragOverEvent) {
-    if (!isDragging) return;
-
-    const activeIdRaw = event.active.id as string;
-    const overIdRaw = event.over?.id as string;
-    if (!overIdRaw) return;
-
-    const activeId = activeIdRaw.replace("task:", "");
-    const isOverIsCol = overIdRaw.startsWith("col:");
-    const overId = isOverIsCol
-      ? overIdRaw.replace("col:", "")
-      : overIdRaw.replace("task:", "");
-
-    const currentTask = optimisticTasks.find((task) => task.id === activeId);
-    if (!currentTask) return;
-
-    if (isOverIsCol) {
-      if (currentTask.status !== overId) {
-        applyLocalStatusChange(activeId, overId as Task["status"]);
-      }
-      return;
-    }
-
-    const overTask = optimisticTasks.find((task) => task.id === overId);
-    if (!overTask) return;
-
-    if (currentTask.status === overTask.status) {
-      const tasksInColumn = tasksByStatus[currentTask.status] || [];
-      const activeIndex = tasksInColumn.findIndex(
-        (task) => task.id === activeId,
-      );
-      const overIndex = tasksInColumn.findIndex((task) => task.id === overId);
-
-      if (activeIndex !== overIndex) {
-        const reorderedTasks = arrayMove(tasksInColumn, activeIndex, overIndex);
-        const updatedTasks = optimisticTasks.map((task) => {
-          if (task.status === currentTask.status) {
-            const newIndex = reorderedTasks.findIndex((t) => t.id === task.id);
-            return newIndex !== -1 ? reorderedTasks[newIndex] : task;
-          }
-          return task;
-        });
-
-        setOptimisticTasks(updatedTasks);
-      }
-    } else {
-      applyLocalStatusChange(activeId, overTask.status);
-    }
-  }
-
-  function handleDragEnd(event: DragEndEvent) {
-    const { active, over } = event;
-    setActiveTask(null);
-    setIsDragging(false);
-    if (!over) return;
-
-    const activeIdRaw = active.id as string;
-    const overIdRaw = over.id as string;
-    const activeTaskId = activeIdRaw.replace("task:", "");
-    const isOverIsCol = overIdRaw.startsWith("col:");
-    const overId = isOverIsCol
-      ? overIdRaw.replace("col:", "")
-      : overIdRaw.replace("task:", "");
-
-    if (pendingUpdatesRef.current.has(activeTaskId)) return;
-
-    let newStatus: Task["status"] | null = null;
-    if (isOverIsCol) {
-      newStatus = overId as Task["status"];
-    } else {
-      const overTask = ((tasksData?.tasks as Task[]) || []).find(
-        (t) => t.id === overId,
-      );
-      newStatus = overTask?.status ?? null;
-    }
-
-    if (!newStatus) return;
-
-    const currentTaskOnServer = ((tasksData?.tasks as Task[]) || []).find(
-      (t) => t.id === activeTaskId,
-    );
-    if (!currentTaskOnServer) return;
-    if (currentTaskOnServer.status === newStatus) return;
-
-    const currentData = active.data.current?.task;
-
-    const tasksInColumn = optimisticTasks.filter(
-      (task) => task.status === newStatus,
-    );
-
-    const newPosition = tasksInColumn.findIndex(
-      (task) => task.id === activeTaskId,
-    );
-
-    beginUpdateSnapshot();
-    addPending(activeTaskId);
-    lastTaskIdRef.current = activeTaskId;
-    applyLocalStatusChange(activeTaskId, newStatus);
-
-    updateTask({
-      variables: { input: { id: activeTaskId, status: newStatus } },
-      optimisticResponse: {
-        __typename: "Mutation",
-        updateTask: {
-          __typename: "Task",
-          id: activeTaskId,
-          status: newStatus,
-          name: currentData.name ?? "",
-          assignee: currentData.assignee ?? null,
-          dueDate: currentData.dueDate ?? null,
-          pointEstimate: currentData.pointEstimate ?? null,
-          tags: currentData.tags ?? [],
-          position: newPosition,
-        },
-      },
-
-      update: (cache) => {
-        cache.modify({
-          id: cache.identify({ __typename: "Task", id: activeTaskId }),
-          fields: {
-            status: () => newStatus,
-          },
-        });
-      },
-    });
-  }
-
-  const sensors = useSensors(
-    useSensor(PointerSensor, {
-      activationConstraint: {
-        distance: 1,
-      },
-    }),
-    useSensor(KeyboardSensor),
-  );
 
   if (isloadingCols || isloadingTasks) return <LoadingComponent />;
 
@@ -331,7 +131,7 @@ export const Tasks = () => {
           <TaskColumn
             key={col.name}
             colName={col.name}
-            tasks={tasksByStatus[col.name] || []}
+            tasks={finalTasksByStatus[col.name] || []}
           />
         ))}
